feat(web): bookmarkable URL routing for dashboard repo views
Extract DashboardShell from page.tsx and add /dashboard/[owner]/[repo]
dynamic route so repo-focused views are bookmarkable and shareable.

- Sidebar repos now use Link navigation instead of onClick state
- Clicking selected repo navigates back to /dashboard (deselect)
- Brand header links back to /dashboard overview
- Right-click "Open in new tab" works on repo rows
