+++ web/src/app/dashboard/components/sidebar.module.css
@@ -77,6 +77,7 @@
 	text-align: left;
 	font: inherit;
 	color: inherit;
+	text-decoration: none;
 	transition: all 0.15s ease;
 }
